refactor PTR component
