+++ app/src/components/QueryTimeline.tsx
@@ -1,18 +1,17 @@
 import { Box, Divider, ListItem, ListItemIcon, ListItemText, type SxProps, Typography, useTheme } from '@mui/material'
-import React, { memo, useCallback, useEffect, useState, useRef, forwardRef, type ForwardedRef } from 'react'
+import React, { memo, useEffect, useState, useRef, forwardRef, type ForwardedRef } from 'react'
 import { AssociationFrame } from './Association/AssociationFrame'
 import { Loading } from './ui/Loading'
 import { MessageContainer } from './Message/MessageContainer'
 import { ErrorBoundary, type FallbackProps } from 'react-error-boundary'
 import HeartBrokenIcon from '@mui/icons-material/HeartBroken'
-import ArrowUpwardIcon from '@mui/icons-material/ArrowUpward'
-import SyncIcon from '@mui/icons-material/Sync'
 import { type Query, type QueryTimelineReader as CoreTimelineReader } from '@concrnt/client'
 import useSound from 'use-sound'
 import { usePreference } from '../context/PreferenceContext'
 import { VList, type VListHandle } from 'virtua'
 import { useClient } from '../context/ClientContext'
 import { UseSoundFormats } from '../constants'
+import { PullToRefresh } from './PullToRefresh'
 
 export interface TimelineProps {
     timeline: string
@@ -22,8 +21,6 @@ export interface TimelineProps {
     header?: JSX.Element
     onScroll?: (top: number) => void
 }
-
-const PTR_HEIGHT = 60
 
 const divider = (
     <Divider
@@ -48,10 +45,7 @@ const timeline = forwardRef((props: TimelineProps, ref: ForwardedRef<VListHandle
     const [hasMoreData, setHasMoreData] = useState<boolean>(false)
     const [isFetching, setIsFetching] = useState<boolean>(false)
 
-    const [touchPosition, setTouchPosition] = useState<number>(0)
-    const [loaderSize, setLoaderSize] = useState<number>(0)
-    const [loadable, setLoadable] = useState<boolean>(false)
-    const [ptrEnabled, setPtrEnabled] = useState<boolean>(false)
+    const positionRef = useRef<number>(0)
 
     const [playBubble] = useSound(sound.post, {
         volume: sound.volume / 100,
@@ -87,58 +81,6 @@ const timeline = forwardRef((props: TimelineProps, ref: ForwardedRef<VListHandle
         }
     }, [props.timeline, props.query, client])
 
-    const positionRef = useRef<number>(0)
-    const scrollParentRef = useRef<HTMLDivElement>(null)
-
-    const onTouchStart = useCallback((raw: Event) => {
-        const e = raw as TouchEvent
-        setTouchPosition(e.touches[0].clientY)
-        setLoadable(positionRef.current === 0)
-    }, [])
-
-    const onTouchMove = useCallback(
-        (raw: Event) => {
-            if (!loadable) return
-            const e = raw as TouchEvent
-            if (!scrollParentRef.current) return
-            const delta = e.touches[0].clientY - touchPosition
-            setLoaderSize(Math.min(Math.max(delta, 0), PTR_HEIGHT))
-            if (delta >= PTR_HEIGHT) setPtrEnabled(true)
-        },
-        [scrollParentRef.current, touchPosition]
-    )
-
-    const onTouchEnd = useCallback(() => {
-        setLoaderSize(0)
-        if (ptrEnabled) {
-            if (isFetching) {
-                setPtrEnabled(false)
-                return
-            }
-            setIsFetching(true)
-            setHasMoreData(false)
-            setTimeout(() => {
-                timeline.current?.reload().then((hasMore) => {
-                    setHasMoreData(hasMore)
-                    setIsFetching(false)
-                    setPtrEnabled(false)
-                })
-            }, 1000)
-        }
-    }, [ptrEnabled, setPtrEnabled, props.timeline, props.query, client.api, isFetching])
-
-    useEffect(() => {
-        if (!scrollParentRef.current) return
-        scrollParentRef.current.addEventListener('touchstart', onTouchStart)
-        scrollParentRef.current.addEventListener('touchmove', onTouchMove)
-        scrollParentRef.current.addEventListener('touchend', onTouchEnd)
-        return () => {
-            scrollParentRef.current?.removeEventListener('touchstart', onTouchStart)
-            scrollParentRef.current?.removeEventListener('touchmove', onTouchMove)
-            scrollParentRef.current?.removeEventListener('touchend', onTouchEnd)
-        }
-    }, [scrollParentRef.current, onTouchStart, onTouchMove, onTouchEnd])
-
     const count = timeline.current?.body.length ?? 0
     let alreadyFetchInThisRender = false
 
@@ -159,147 +101,106 @@ const timeline = forwardRef((props: TimelineProps, ref: ForwardedRef<VListHandle
             })
     }
 
-    return (
-        <>
-            <Box
-                sx={{
-                    height: `${ptrEnabled ? PTR_HEIGHT : loaderSize}px`,
-                    width: '100%',
-                    position: 'relative',
-                    color: 'text.secondary',
-                    display: 'flex',
-                    transition: 'height 0.2s ease-in-out',
-                    overflow: 'hidden'
-                }}
-            >
-                <Box
-                    display="flex"
-                    alignItems="center"
-                    justifyContent="center"
-                    height={`${PTR_HEIGHT}px`}
-                    position="absolute"
-                    width="100%"
-                    bottom="0"
-                    left="0"
-                >
-                    {isFetching ? (
-                        <SyncIcon
-                            sx={{
-                                animation: 'spin 1s linear infinite',
-                                '@keyframes spin': {
-                                    '0%': { transform: 'rotate(0deg)' },
-                                    '100%': { transform: 'rotate(-360deg)' }
-                                }
-                            }}
-                        />
-                    ) : (
-                        <ArrowUpwardIcon
-                            sx={{
-                                transform: `rotate(${ptrEnabled ? 0 : 180}deg)`,
-                                transition: 'transform 0.2s ease-in-out'
-                            }}
-                        />
-                    )}
-                </Box>
-            </Box>
-            <Box
-                ref={scrollParentRef}
-                sx={{
-                    display: 'flex',
-                    flex: 1,
-                    overflow: 'hidden',
-                    flexDirection: 'column'
-                }}
-            >
-                <VList
-                    style={{
-                        flex: 1,
-                        display: 'flex',
-                        flexDirection: 'column',
-                        listStyle: 'none',
-                        overflowX: 'hidden',
-                        overflowY: 'auto',
-                        overscrollBehaviorY: 'none',
-                        scrollbarGutter: 'stable'
-                    }}
-                    onScroll={(top) => {
-                        positionRef.current = top
-                        props.onScroll?.(top)
-                    }}
-                    onRangeChange={(_, end) => {
-                        if (end + 3 > count && hasMoreData) readMore()
-                    }}
-                    ref={ref}
-                >
-                    {props.header}
+    const onRefresh = async (): Promise<void> => {
+        console.log('refresh!')
+        setIsFetching(true)
+        setHasMoreData(false)
+        const hasMore = (await timeline.current?.reload()) ?? false
+        setHasMoreData(hasMore)
+        await new Promise((resolve) => setTimeout(resolve, 1000))
+        setIsFetching(false)
+    }
 
-                    {(timeline.current?.body.length ?? 0) === 0
-                        ? !timelineLoading && (
-                              <Box
+    return (
+        <PullToRefresh positionRef={positionRef} isFetching={isFetching} onRefresh={onRefresh}>
+            <VList
+                style={{
+                    flex: 1,
+                    display: 'flex',
+                    flexDirection: 'column',
+                    listStyle: 'none',
+                    overflowX: 'hidden',
+                    overflowY: 'auto',
+                    overscrollBehaviorY: 'none',
+                    scrollbarGutter: 'stable'
+                }}
+                onScroll={(top) => {
+                    positionRef.current = top
+                    props.onScroll?.(top)
+                }}
+                onRangeChange={(_, end) => {
+                    if (end + 3 > count && hasMoreData) readMore()
+                }}
+                ref={ref}
+            >
+                {props.header}
+
+                {(timeline.current?.body.length ?? 0) === 0
+                    ? !timelineLoading && (
+                          <Box
+                              sx={{
+                                  width: '100%',
+                                  display: 'flex',
+                                  justifyContent: 'center'
+                              }}
+                          >
+                              <Typography
                                   sx={{
-                                      width: '100%',
-                                      display: 'flex',
-                                      justifyContent: 'center'
+                                      color: 'text.secondary'
                                   }}
                               >
-                                  <Typography
-                                      sx={{
-                                          color: 'text.secondary'
-                                      }}
-                                  >
-                                      No currents yet here!
-                                  </Typography>
-                              </Box>
+                                  No currents yet here!
+                              </Typography>
+                          </Box>
+                      )
+                    : (timelineLoading ? [] : (timeline.current?.body ?? [])).map((e) => {
+                          let element
+                          const type = e.resourceID[0]
+                          switch (type) {
+                              case 'm':
+                                  element = (
+                                      <MessageContainer
+                                          dimOnHover
+                                          sx={timelineElemSx}
+                                          messageID={e.resourceID}
+                                          messageOwner={e.owner}
+                                          resolveHint={e.timelineID.split('@')[1]}
+                                          lastUpdated={e.lastUpdate?.getTime() ?? 0}
+                                          after={divider}
+                                          timestamp={e.lastUpdate}
+                                      />
+                                  )
+                                  break
+                              case 'a':
+                                  element = (
+                                      <AssociationFrame
+                                          dimOnHover
+                                          sx={timelineElemSx}
+                                          associationID={e.resourceID}
+                                          associationOwner={e.owner}
+                                          lastUpdated={e.lastUpdate?.getTime() ?? 0}
+                                          after={divider}
+                                          perspective={props.perspective}
+                                      />
+                                  )
+                                  break
+                              default:
+                                  element = <Typography>Unknown message type: {type}</Typography>
+                                  break
+                          }
+
+                          return (
+                              <React.Fragment key={e.resourceID}>
+                                  <ErrorBoundary FallbackComponent={renderError}>{element}</ErrorBoundary>
+                              </React.Fragment>
                           )
-                        : (timelineLoading ? [] : (timeline.current?.body ?? [])).map((e) => {
-                              let element
-                              const type = e.resourceID[0]
-                              switch (type) {
-                                  case 'm':
-                                      element = (
-                                          <MessageContainer
-                                              dimOnHover
-                                              sx={timelineElemSx}
-                                              messageID={e.resourceID}
-                                              messageOwner={e.owner}
-                                              resolveHint={e.timelineID.split('@')[1]}
-                                              lastUpdated={e.lastUpdate?.getTime() ?? 0}
-                                              after={divider}
-                                              timestamp={e.lastUpdate}
-                                          />
-                                      )
-                                      break
-                                  case 'a':
-                                      element = (
-                                          <AssociationFrame
-                                              dimOnHover
-                                              sx={timelineElemSx}
-                                              associationID={e.resourceID}
-                                              associationOwner={e.owner}
-                                              lastUpdated={e.lastUpdate?.getTime() ?? 0}
-                                              after={divider}
-                                              perspective={props.perspective}
-                                          />
-                                      )
-                                      break
-                                  default:
-                                      element = <Typography>Unknown message type: {type}</Typography>
-                                      break
-                              }
+                      })}
 
-                              return (
-                                  <React.Fragment key={e.resourceID}>
-                                      <ErrorBoundary FallbackComponent={renderError}>{element}</ErrorBoundary>
-                                  </React.Fragment>
-                              )
-                          })}
-
-                    {!timelineLoading && isFetching && (
-                        <Loading key={0} message="Loading..." color={theme.palette.text.primary} />
-                    )}
-                </VList>
-            </Box>
-        </>
+                {!timelineLoading && isFetching && (
+                    <Loading key={0} message="Loading..." color={theme.palette.text.primary} />
+                )}
+            </VList>
+        </PullToRefresh>
     )
 })
 timeline.displayName = 'timeline'
